remove validation, move files inside update folders, update OeNumber to OeNumberKey
