+++ src/controllers/index.ts
@@ -7,8 +7,8 @@ export * as beneficialOwnerStatements from "./beneficial.owner.statements.contro
 export * as beneficialOwnerType from "./beneficial.owner.type.controller";
 export * as checkYourAnswers from "./check.your.answers.controller";
 export * as landing from "./landing.controller";
-export * as updateLanding from "./update.landing.controller";
-export * as overseasEntityQuery from "./overseas.entity.query.controller";
+export * as updateLanding from "./update/update.landing.controller";
+export * as overseasEntityQuery from "./update/overseas.entity.query.controller";
 export * as managingOfficerIndividual from "./managing.officer.controller";
 export * as managingOfficerCorporate from "./managing.officer.corporate.controller";
 export * as presenter from "./presenter.controller";
