+++ src/controllers/update/overseas.entity.query.controller.ts
@@ -1,9 +1,9 @@
 import { NextFunction, Request, Response } from "express";
 
-import { logger } from "../utils/logger";
-import * as config from "../config";
-import { getApplicationData, setExtraData } from "../utils/application.data";
-import { OeNumber } from "../model/data.types.model";
+import { logger } from "../../utils/logger";
+import * as config from "../../config";
+import { getApplicationData, setExtraData } from "../../utils/application.data";
+import { OeNumberKey } from "../../model/data.types.model";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
   try {
@@ -22,14 +22,11 @@ export const get = (req: Request, res: Response, next: NextFunction) => {
 export const post = (req: Request, res: Response, next: NextFunction) => {
   try {
     logger.debugRequest(req, `POST ${config.OVERSEAS_ENTITY_QUERY_PAGE}`);
-    const oeNumber = req.body[OeNumber];
+    const oeNumber = req.body[OeNumberKey];
 
-    if ( oeNumber.length === 8 ) {
-      setExtraData(req.session, { ...getApplicationData(req.session), [OeNumber]: oeNumber });
-      return res.redirect(config.CONFIRM_OVERSEAS_COMPANY_PROFILES_URL);
-    } else {
-      return res.render(config.OVERSEAS_ENTITY_QUERY_PAGE);
-    }
+    setExtraData(req.session, { ...getApplicationData(req.session), [OeNumberKey]: oeNumber });
+    return res.redirect(config.CONFIRM_OVERSEAS_COMPANY_PROFILES_URL);
+
   } catch (error) {
     logger.errorRequest(req, error);
     next(error);
