+++ src/controllers/update/update.landing.controller.ts
@@ -1,7 +1,7 @@
 import { NextFunction, Request, Response } from "express";
 
-import { logger } from "../utils/logger";
-import * as config from "../config";
+import { logger } from "../../utils/logger";
+import * as config from "../../config";
 
 export const get = (req: Request, res: Response, next: NextFunction) => {
   try {
