+++ test/controllers/update/update.landing.controller.spec.ts
@@ -1,11 +1,11 @@
 jest.mock("ioredis");
-jest.mock("../../src/utils/application.data");
+jest.mock("../../../src/utils/application.data");
 
 import { describe, expect, test, jest } from '@jest/globals';
 import request from "supertest";
 
-import app from "../../src/app";
-import { UPDATE_LANDING_PAGE_URL, UPDATE_LANDING_URL } from "../../src/config";
+import app from "../../../src/app";
+import { UPDATE_LANDING_PAGE_URL, UPDATE_LANDING_URL } from "../../../src/config";
 
 describe("UPDATE LANDING controller", () => {
